requirements.txt
playwright>=1.43.0
pandas>=2.0.3
tabulate>=0.9.0
jinja2>=3.1.6
Pygments==2.15.1
PyYAML==6.0
pygments-better-html==0.1.4
ruamel.yaml>=0.18.6
chardet>=5.2.0
colorama>=0.4.6
fastapi>=0.116.1
uvicorn[standard]>=0.29.0
SQLAlchemy>=2.0.35
pydantic>=2.8.2
